% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran_test.R
\name{local_moran}
\alias{local_moran}
\title{Calculate the Local Moran's I Statistic}
\usage{
local_moran(x, neighbors, weights, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{neighbors}{A neighbor list object as created by \code{st_neighbors()}.}

\item{weights}{A weights list as created by \code{st_weights()}.}

\item{...}{See \code{?spdep::localmoran()} for more options.}
}
\description{
Calculate the Local Moran's I Statistic
}
