% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran_test.R
\name{moran_mc}
\alias{moran_mc}
\title{Moran's I Permutation Test}
\usage{
moran_mc(data, x, nb, wt, nsim = 999, alternative = "greater", ...)
}
\description{
Moran's I Permutation Test
}
\seealso{
Other stats: 
\code{\link{geary_test}()},
\code{\link{local_moran}()},
\code{\link{moran_test}()},
\code{\link{st_lag}()},
\code{\link{st_nb_dists}()}
}
\concept{stats}
