% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_match_test.R
\name{neighbor_match_test}
\alias{neighbor_match_test}
\alias{neighbor_match_test.default}
\alias{neighbor_match_test.data.frame}
\title{Local Neighbor Match Test}
\usage{
neighbor_match_test(x, ...)

\method{neighbor_match_test}{default}(
  .cols,
  nb,
  k = 10,
  method = "euclidean",
  scale = TRUE,
  p = 2
)

\method{neighbor_match_test}{data.frame}(
  .data,
  .cols,
  nb,
  k = 10,
  method = "euclidean",
  scale = TRUE,
  p = 2
)
}
\arguments{
\item{.cols}{A list of numeric vectors. Intended to be used inside of a \code{dplyr::mutate()} call. In the case of a data.frame, the columns to be used in calculating the distance matrix using tidy selection.}

\item{nb}{The unquoted name of your knn neighbor list.}

\item{k}{The number of neighbors to identify. Should be the same \code{k} used in \code{st_knn()}.}

\item{method}{The distance measure to be used by \code{stats::dist()}.}

\item{scale}{Whether or not to standardize columns prior to calculation. It is strongly recommended to do so.}

\item{p}{The power of the Minkowski distance.}

\item{.data}{A data frame.}
}
\description{
Implements the Local Neighbor Match Test as described in \emph{Tobler's Law in a Multivariate World} (Anselin and Li, 2020).
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Conduct a neighbor match test within a mutate call.

\item \code{data.frame}: Conduct a neighbor match test with a data frame.
}}

