% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_weights.R
\name{st_inverse_weights}
\alias{st_inverse_weights}
\title{Calculate Inverse Distance Bands}
\usage{
st_inverse_weights(x, knn)
}
\arguments{
\item{x}{Spatial points. Typically the \code{geometry} column of an sf object.}

\item{knn}{A knn neighbor list as created from \code{st_knn()}.}
}
\description{
Calculate Inverse Distance Bands
}
\details{
See implementation details \href{https://spatialanalysis.github.io/lab_tutorials/Spatial_Weights_as_Distance_Functions.html#kernal-weights}{here}.
}
