% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_weights.R
\name{st_kernal_weight}
\alias{st_kernal_weight}
\title{Calculate Kernal Weights}
\usage{
st_kernal_weight(x, knn, kernal = "uniform")
}
\arguments{
\item{x}{Spatial points. Typically the \code{geometry} column of an sf object.}

\item{knn}{A knn neighbor list as created from \code{st_knn()}.}

\item{kernal}{One of "uniform", "gaussian",  "triangular", "epanechnikov", or "quartic".}
}
\description{
Calculate Kernal Weights
}
