% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_weights.R
\name{st_kernel_weight}
\alias{st_kernel_weight}
\title{Calculate Kernel Weights}
\usage{
st_kernel_weight(x, knn, kernel = "uniform")
}
\arguments{
\item{x}{Spatial points. Typically the \code{geometry} column of an sf object.}

\item{knn}{A knn neighbor list as created from \code{st_knn()}.}

\item{kernel}{One of "uniform", "gaussian",  "triangular", "epanechnikov", or "quartic".}
}
\description{
Calculate Kernel Weights
}
