% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_knn.R
\name{st_knn}
\alias{st_knn}
\title{Calculate K-Nearest Neighbors}
\usage{
st_knn(x, k = 1, ...)
}
\arguments{
\item{x}{An sf or sfc object.}
}
\description{
Calculate K-Nearest Neighbors
}
\seealso{
Other neighbors: 
\code{\link{st_contiguity}()},
\code{\link{st_nb_band}()}
}
\concept{neighbors}
