% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_lag.R
\name{st_lag}
\alias{st_lag}
\title{Calculate spatial lag}
\usage{
st_lag(x, nb, wt, allow_zero = NULL, na_ok = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{nb}{A neighbor list object as created by \code{st_neighbors()}.}

\item{wt}{A weights list as created by \code{st_weights()}.}

\item{allow_zero}{If \code{TRUE}, assigns zero as lagged value to zone without neighbors.}

\item{na_ok}{Default \code{FALSE}. If, \code{TRUE} missing values return a lagged \code{NA}.}

\item{...}{See \code{?spdep::lag.listw} for more.}
}
\description{
Calculate spatial lag
}
\seealso{
Other stats: 
\code{\link{geary_test}()},
\code{\link{local_moran}()},
\code{\link{moran_mc}()},
\code{\link{moran_test}()},
\code{\link{st_nb_dists}()}
}
\concept{stats}
