% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_nb_dist.R
\name{st_nb_dist}
\alias{st_nb_dist}
\title{Calculate neighbor distances}
\usage{
st_nb_dist(x, nb, longlat = NULL)
}
\arguments{
\item{x}{Spatial points. Typically the \code{geometry} column of an sf object.}

\item{longlat}{\code{TRUE} if point coordinates are longitude-latitude decimal degrees, in which case distances are measured in kilometers. See \code{?spdep::nbdists()} for more.}
}
\description{
From an nb list and point geometry, return a list of distances for each observation's neighbors list.
}
\details{
Utilizes \code{spdep::nbdists()} for distance calculation.
}
