% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_neighbors.R
\name{st_nb_lag}
\alias{st_nb_lag}
\title{Pure Higher Order Neighbors}
\usage{
st_nb_lag(nb, order)
}
\arguments{
\item{nb}{A neighbor list object as created by \code{st_contiguity()}.}

\item{order}{The order of neighbors.}
}
\description{
Pure Higher Order Neighbors
}
\seealso{
Other other: 
\code{\link{categorize_lisa}()},
\code{\link{st_cardinalties}()},
\code{\link{st_nb_lag_cumul}()}
}
\concept{other}
