% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_neighbors.R
\name{st_neighbor_lag}
\alias{st_neighbor_lag}
\title{Pure Higher Order Neighbors}
\usage{
st_neighbor_lag(neighbors, order)
}
\arguments{
\item{neighbors}{A neighbor list object as created by \code{st_neighbors()}.}

\item{order}{The order of neighbors.}
}
\description{
Pure Higher Order Neighbors
}
