% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_neighbors.R
\name{st_neighbor_lag_cumul}
\alias{st_neighbor_lag_cumul}
\title{Encompassing Higher Order Neighbors}
\usage{
st_neighbor_lag_cumul(neighbors, order)
}
\arguments{
\item{neighbors}{A neighbor list object as created by \code{st_neighbors()}.}

\item{order}{The order of neighbors.}
}
\description{
Creates an encompassing neighbor list of the order specified.
For example, if the order is 2 the result contains both 1st
and 2nd order neighbors.
}
