% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_neighbors.R
\name{st_neighbors}
\alias{st_neighbors}
\title{Identify polygon neighbors}
\usage{
st_neighbors(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}. Use \code{st_contiguity} instead. This function creates neighbors based on contiguities.
}
\keyword{internal}
