% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran_test.R
\name{unpack_lisa}
\alias{unpack_lisa}
\title{Unpack LISA data frame column}
\usage{
unpack_lisa(x, lisa, ...)
}
\arguments{
\item{x}{An sf object}

\item{lisa}{The unquoted column name containing the LISA data frame}
}
\description{
Unpack LISA data frame column
}
