% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran_test.R
\name{categorize_lisa}
\alias{categorize_lisa}
\title{Categorize LISA}
\usage{
categorize_lisa(x, x_lag, scale = TRUE)
}
\arguments{
\item{x}{Numeric vector.}

\item{x_lag}{The spatial lag of x as calculated by \code{st_lag()}.}
}
\description{
Categorize LISA
}
