% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geary.R
\name{geary_test}
\alias{geary_test}
\title{Conducts a Geary's C test}
\usage{
geary_test(x, neighbors, weights, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{neighbors}{A neighbor list object as created by \code{st_neighbors()}.}

\item{weights}{A weights list as created by \code{st_weights()}.}
}
\description{
Conducts a Geary's C test
}
\seealso{
Other stats: 
\code{\link{local_moran}()},
\code{\link{moran_mc}()},
\code{\link{moran_test}()},
\code{\link{st_lag}()},
\code{\link{st_nb_dists}()}
}
\concept{stats}
