% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran_test.R
\name{moran_test}
\alias{moran_test}
\title{Conduct a Moran's I test}
\usage{
moran_test(data, x, neighbors, weights, ...)
}
\description{
Conduct a Moran's I test
}
\examples{
acs \%>\%
  mutate(nb = st_neighbors(geometry),
       wt = st_weights(nb)) \%>\%
  moran_test(bach, nb, wt)
}
