% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_cardinalities.R
\name{st_cardinalties}
\alias{st_cardinalties}
\title{Calculate neighbor cardinalities}
\usage{
st_cardinalties(neighbors)
}
\arguments{
\item{neighbors}{A neighbor list object as created by \code{st_neighbors()}.}
}
\description{
Calculate neighbor cardinalities
}
\details{
See \code{?spdep::card()} for more.
}
\seealso{
Other other: 
\code{\link{categorize_lisa}()},
\code{\link{st_nb_lag_cumul}()},
\code{\link{st_nb_lag}()}
}
\concept{other}
