% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_neighbors.R
\name{st_neighbors}
\alias{st_neighbors}
\title{Identify polygon neighbors}
\usage{
st_neighbors(x, queen = TRUE, ...)
}
\arguments{
\item{x}{An sf or sfc object.}

\item{queen}{Default \code{TRUE}. For more see \code{?spdep::poly2nb}}
}
\description{
Identify polygon neighbors
}
